Major Refractor
Updated the bot to have better readability and added some new functions.
The bot now has a class called DiscordClock which stores all relavent
info to a clock-in/out. The bot also has the ability to tell if a
discord clock is valid or not by using the DiscordClock's is_valid
variable. It will let the end user know if the clock the entered is
invalid by a reaction and sending a message with the reason it is
invalid.